Make ocr tool run recursively on directory

+++ tools/ocr.js
@@ -7,12 +7,41 @@ function base64Encode(file) {
     return new Buffer(bitmap).toString('base64');
 }
 
-const inputFile = process.argv[2];
+function walkDir(dir, callback) {
+    fs.readdirSync(dir).forEach( f => {
+        let dirPath = path.join(dir, f);
+        let isDirectory = fs.statSync(dirPath).isDirectory();
+        isDirectory ? 
+            walkDir(dirPath, callback) : callback(path.join(dir, f));
+    });
+};
+
+function getOutputFile(inputFile, outputDir) {
+    return path.join(outputDir, path.basename(inputFile) + ".google.json");
+}
+
+function ocrFile(inputFile, outputFile, key) {
+    if (fs.existsSync(outputFile)) {
+        console.log(`${outputFile} already exists`);
+    } else {
+        const base64EncodedFileContent = base64Encode(inputFile);
+        ocr.makeVisionRequest(base64EncodedFileContent, key,
+            response => fs.writeFileSync(outputFile, JSON.stringify(response, null, 2)));
+    }
+}
+
+const input = process.argv[2];
 const outputDir = process.argv[3];
 const key = process.argv[4];
 
-const base64EncodedFileContent = base64Encode(inputFile);
-const outputFile = path.join(outputDir, path.basename(inputFile) + ".google.json")
-
-ocr.makeVisionRequest(base64EncodedFileContent, key,
-    response => fs.writeFileSync(outputFile, JSON.stringify(response, null, 2)));
+if (fs.lstatSync(input).isDirectory()) {
+    // walk directory and run for each file
+    walkDir(input, function(inputFile) {
+        if (inputFile.toLowerCase().endsWith('.jpg')) {
+            ocrFile(inputFile, getOutputFile(inputFile, outputDir), key);
+        }
+    });
+} else {
+    // single file
+    ocrFile(input, getOutputFile(input, outputDir), key);
+}
